Add viles via upload
This file builds a data pipline for appl data. 

It then creates the correct indexing and frequency for time series analysis. 

Following this it brings in rolling averages for analyzing the data. 

Splits the data and assigns it to the proper variables before visualizing the data on a chart to ensure effective methods. 

A linear regression using rolling averages as the features is then utilized. 

MAE, MSE, and RMSE are used to analyze the results. 

The results are then visualized via plot and then the forecasted values are compared against the actual values. The implications are then analyzed. 

Next, exponential smoothing is used on a linear regression model. 

3 periods are forecasted and the data is compared vs the forecast. 

Same error scores as above are utilized. 

Lastly, an LSTM model with data scaling is used from Keras. 

Then tensorflow and kerastuners are used to optimize the hyperparameters of the models. 

It is then visualized with a graph and comparatively analyzed.