Files added via upload
Attached is a customer segmentation model. It takes the provided datafile, via file_path to then be used for later methods. 

The first thing done is data exploration. This is done to help understand the distribution of the data sources, observe trends, and understand what level of pre-processing, if any, needs to be done. 

Next, enhanced clustering occurs with the K-means method. To evaluate the model the elbow method & silhouette score are used. 

Then advanced hierarchical clustering with a dendrogram is used. The silhouette score, DBI, and CHI score are then used. 

They are then compared using a bar graph. 

Then a DBSCAN clustering is utilized. The silhouette, DBI, and CHI score are used to optimize the model with CHI outperforming by a large amount. 

Lastly, a PCA is applied and helps to explain the data better.