Added via upload
This project is one of my favorites. 

It utilizes natural language processing to understand user problems, and map the correct statistical technique to use based on the training data. That method can then be copy and pasted for use in the deployer method. 

This project is about 80% of the way to a Statistical Action Model, which would help entry-level data scientists rapidly advance their time-to-profitability as a resource for an institution for a bank, or research center of excellence.